docs: fix Sandpack component
