+++ docs/package.json
@@ -20,6 +20,7 @@
     "@astrojs/tailwind": "^2.1.1",
     "@astropub/md": "^0.2.0",
     "@codesandbox/sandpack-react": "^2.6.2",
+    "@codesandbox/sandpack-themes": "^2.0.21",
     "@docsearch/css": "^3.3.0",
     "@docsearch/react": "^3.3.0",
     "@types/react": "^18.0.23",
