+++ docs/src/pages/en/components/media-play-button.mdx
@@ -5,7 +5,7 @@ layout: ../../../layouts/ComponentLayout.astro
 source: https://github.com/muxinc/media-chrome/tree/main/src/js/media-play-button.js
 ---
 
-import ComponentSandpack from "../../../components/ComponentSandpack";
+import SandpackContainer from "../../../components/SandpackContainer.astro";
 
 The `<media-play-button>` component is used to toggle your media playback state.  
 In simpler terms, it is used to both play and pause your media content.
@@ -17,8 +17,7 @@ once your media playback state changes.
 
 ## Default usage
 
-<ComponentSandpack
-  client:load
+<SandpackContainer
   html={`<media-controller>
   <video
     slot="media"
@@ -39,8 +38,7 @@ the default one provided by media-chrome.
 Here's an example of how you can replace the default play and pause icons with 
 the literal words "Play" and "Pause":  
 
-<ComponentSandpack
-  client:load
+<SandpackContainer
   height={270}
   html={`<media-controller>
   <video
